Refactor check injection into a separate function
