+++ kani-compiler/src/codegen_cprover_gotoc/codegen/place.rs
@@ -8,7 +8,6 @@
 use super::typ::TypeExt;
 use crate::codegen_cprover_gotoc::codegen::ty_stable::pointee_type;
 use crate::codegen_cprover_gotoc::codegen::typ::std_pointee_type;
-use crate::codegen_cprover_gotoc::codegen::PropertyClass;
 use crate::codegen_cprover_gotoc::utils::{dynamic_fat_ptr, slice_fat_ptr};
 use crate::codegen_cprover_gotoc::GotocCtx;
 use crate::unwrap_or_return_codegen_unimplemented;
@@ -17,7 +16,7 @@ use rustc_middle::ty::layout::LayoutOf;
 use rustc_smir::rustc_internal;
 use rustc_target::abi::{TagEncoding, Variants};
 use stable_mir::mir::{FieldIdx, Local, Mutability, Place, ProjectionElem};
-use stable_mir::ty::{RigidTy, Ty, TyKind, TypeAndMut, VariantDef, VariantIdx};
+use stable_mir::ty::{RigidTy, Ty, TyKind, VariantDef, VariantIdx};
 use tracing::{debug, trace, warn};
 
 /// A projection in Kani can either be to a type (the normal case),
@@ -652,16 +651,11 @@ impl<'tcx> GotocCtx<'tcx> {
     ///   build the fat pointer from there.
     /// - For `*(Wrapper<T>)` where `T: Unsized`, the projection's `goto_expr` returns an object,
     ///   and we need to take it's address and build the fat pointer.
-    pub fn codegen_place_ref_stable(
-        &mut self,
-        place: &Place,
-        loc: &Location,
-        inject_check_if_ptr_to_ref_cast: bool,
-    ) -> Expr {
+    pub fn codegen_place_ref_stable(&mut self, place: &Place) -> Expr {
         let place_ty = self.place_ty_stable(place);
         let projection =
             unwrap_or_return_codegen_unimplemented!(self, self.codegen_place_stable(place));
-        let place_ref = if self.use_thin_pointer_stable(place_ty) {
+        if self.use_thin_pointer_stable(place_ty) {
             // For ZST objects rustc does not necessarily generate any actual objects.
             let need_not_be_an_object = self.is_zst_object(&projection.goto_expr);
             let address_of = projection.goto_expr.clone().address_of();
@@ -705,69 +699,7 @@ impl<'tcx> GotocCtx<'tcx> {
                 let len = fat_ptr.member("len", &self.symbol_table);
                 slice_fat_ptr(place_type, data, len, &self.symbol_table)
             }
-        };
-
-        // If converting a raw pointer to a reference, &(*ptr), need to inject
-        // a check to make sure that the pointer points to a valid memory location,
-        // since dereferencing an invalid pointer is UB in Rust.
-        if inject_check_if_ptr_to_ref_cast {
-            if let Some(ProjectionElem::Deref) = place.projection.last() {
-                // Create a place without the topmost dereference projection.
-                let ptr_place = {
-                    let mut ptr_place = place.clone();
-                    ptr_place.projection.pop();
-                    ptr_place
-                };
-                // Only inject the check if dereferencing a raw pointer.
-                let ptr_place_ty = self.place_ty_stable(&ptr_place);
-                if ptr_place_ty.kind().is_raw_ptr() {
-                    // Extract the size of the pointee.
-                    let pointee_size = {
-                        let TypeAndMut { ty: pointee_ty, .. } =
-                            ptr_place_ty.kind().builtin_deref(true).unwrap();
-                        let pointee_ty_layout = pointee_ty.layout().unwrap();
-                        pointee_ty_layout.shape().size.bytes()
-                    };
-
-                    // __CPROVER_r_ok fails if size == 0, so need to explicitly avoid the check.
-                    if pointee_size != 0 {
-                        // Encode __CPROVER_r_ok(ptr, size).
-                        // First, generate a CBMC expression representing the pointer.
-                        let ptr = {
-                            let ptr_projection = self.codegen_place_stable(&ptr_place).unwrap();
-                            if self.use_thin_pointer_stable(place_ty) {
-                                ptr_projection.goto_expr().clone()
-                            } else {
-                                ptr_projection
-                                    .goto_expr()
-                                    .clone()
-                                    .member("data", &self.symbol_table)
-                            }
-                        };
-                        // Then, generate a __CPROVER_r_ok check.
-                        let raw_ptr_read_ok_expr = Expr::read_ok(
-                            ptr.cast_to(Type::void_pointer()),
-                            Expr::int_constant(pointee_size, Type::size_t()),
-                        )
-                        .cast_to(Type::Bool);
-                        // Finally, assert that the pointer points to a valid memory location.
-                        let raw_ptr_read_ok = self.codegen_assert(
-                            raw_ptr_read_ok_expr,
-                            PropertyClass::SafetyCheck,
-                            "dereferencing a pointer to invalid memory location",
-                            *loc,
-                        );
-                        let typ = place_ref.typ().clone();
-                        return Expr::statement_expression(
-                            vec![raw_ptr_read_ok, place_ref.as_stmt(*loc)],
-                            typ,
-                        );
-                    }
-                }
-            }
         }
-        // Otherwise, just return place_ref.
-        place_ref
     }
 
     /// Given a MIR place, generate a CBMC expression that represents it as a CBMC lvalue.
