+++ kani-compiler/src/codegen_cprover_gotoc/codegen/statement.rs
@@ -338,7 +338,7 @@ impl<'tcx> GotocCtx<'tcx> {
             }
             InstanceKind::Shim => {
                 // Since the reference is used right away here, no need to inject a check for pointer validity.
-                let place_ref = self.codegen_place_ref_stable(place, &loc, false);
+                let place_ref = self.codegen_place_ref_stable(place);
                 match place_ty.kind() {
                     TyKind::RigidTy(RigidTy::Dynamic(..)) => {
                         // Virtual drop via a vtable lookup.
